requirements.txt
faiss-cpu
sentence-transformers
numpy
requests